# internal/interface/controller/user/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package gen

import (
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// PostUserRequest defines model for PostUserRequest.
type PostUserRequest struct {
	Birthday string `json:"birthday"`
	Id       string `json:"id"`
	Name     string `json:"name"`
}

// PutUserRequest defines model for PutUserRequest.
type PutUserRequest struct {
	Birthday string `json:"birthday"`
	Name     string `json:"name"`
}

// UserResponse defines model for UserResponse.
type UserResponse struct {
	Birthday string `json:"birthday"`
	Id       string `json:"id"`
	Name     string `json:"name"`
}

// UserIdParameter defines model for UserIdParameter.
type UserIdParameter = openapi_types.UUID

// PostUserJSONRequestBody defines body for PostUser for application/json ContentType.
type PostUserJSONRequestBody = PostUserRequest

// DeleteUserJSONRequestBody defines body for DeleteUser for application/json ContentType.
type DeleteUserJSONRequestBody = PutUserRequest

// PutUserJSONRequestBody defines body for PutUser for application/json ContentType.
type PutUserJSONRequestBody = PutUserRequest